cutting_app/requirements.txt
PyQt6>=6.6
SQLAlchemy>=2.0
numpy>=1.26
openpyxl>=3.1
reportlab>=4.0
Pillow>=10.0
qrcode>=7.4
python-barcode>=0.15
pytest>=7.4
pytest-cov>=4.0